[aerosolve] Drop hive partition before writing to location for external table
