+++ airlearner/airlearner-utils/src/main/scala/com/airbnb/common/ml/util/PipelineUtil.scala
@@ -60,14 +60,20 @@ object PipelineUtil extends ScalaLogging {
     }
   }
 
-  // Save result to output as textfile and update outputTable as partitionSpec
-  // This function only works for external table!
+  /*
+  * Save result to output as textfile and update outputTable as partitionSpec
+  * This function only works for external table!
+  **/
   def saveToHdfsAndUpdateHive[U](hc: HiveContext,
       output: String,
       outputTable: String,
       partitionSpec: String,
       result: RDD[U],
       overwrite: Boolean):Unit = {
+
+    if (overwrite) {
+      HiveUtil.dropHivePartition(hc, outputTable, partitionSpec)
+    }
     PipelineUtil.saveAndCommitAsTextFile(result, output, overwrite)
     HiveUtil.updateHivePartition(hc, outputTable, partitionSpec, output)
   }
@@ -122,7 +128,7 @@ object PipelineUtil extends ScalaLogging {
 
     // Drop the existing partition and write out the
     // DataFrame to Hive in a partition-aware manner
-    dropHivePartition(hc, hiveTableName, hivePartitionSpecs)
+    HiveUtil.dropHivePartition(hc, hiveTableName, hivePartitionSpecs)
     df.write
       .partitionBy(hivePartitionSpecs.keys.toSeq:_*)
       .mode(SaveMode.Append)
@@ -153,51 +159,14 @@ object PipelineUtil extends ScalaLogging {
 
     // Drop the existing partition and write out the
     // DataFrame to Hive in a partition-aware manner
-    dropHivePartition(hc, hiveTableName, hivePartitionSpecs)
+    HiveUtil.dropHivePartition(hc, hiveTableName, hivePartitionSpecs)
     df.write
       .partitionBy(hivePartitionSpecs.keys.toSeq: _*)
       .mode(SaveMode.Append)
       .saveAsTable(hiveTableName)
   }
 
-  /*
-   * Drop the specified Hive partition if it exists
-   */
-  def dropHivePartition(
-      hc: HiveContext,
-      hiveTableName: String,
-      hivePartitionSpecs: Map[String, Any]
-  ): Unit = {
-    if (!hiveTableName.contains('.')) {
-      throw new RuntimeException(s"Missing namespace for hive table: $hiveTableName.")
-    }
 
-    // Break the table name into namespace.table_name
-    val Array(namespace, table) = hiveTableName.split('.')
-    // Turn the partition spec map into a Hive-format String
-    val partitionSpec = hivePartitionSpecsMapToString(hivePartitionSpecs)
-
-    // Drop the partition if it exists
-    hc.sql(s"USE $namespace")
-    hc.sql(s"ALTER TABLE $table DROP IF EXISTS PARTITION ($partitionSpec)")
-  }
-
-  /*
-   * Convert a partition_specs map to a usable Hive string.
-   * ex:
-   *   Map("ds" -> "2016-11-12", "night" -> 3)
-   *   =>
-   *   "ds='2016-11-12', night=3"
-   */
-  private def hivePartitionSpecsMapToString(hivePartitionSpecs: Map[String, Any]): String = {
-    hivePartitionSpecs
-      .map {
-        // If the value is string type, surround it by single-quotes
-        case (column, value: String) => s"$column='$value'"
-        case (column, value) => s"$column=$value"
-      }
-      .mkString(", ")
-  }
 
   def getLastPartition(hc: HiveContext, table: String): String = {
     val query = "SHOW PARTITIONS %s".format(table)
